+++ pkg/services/publicdashboards/api/common_test.go
@@ -16,8 +16,8 @@ import (
 	"github.com/grafana/grafana/pkg/infra/log"
 	"github.com/grafana/grafana/pkg/models"
 	"github.com/grafana/grafana/pkg/plugins"
+	"github.com/grafana/grafana/pkg/services/accesscontrol/acimpl"
 	"github.com/grafana/grafana/pkg/services/accesscontrol/database"
-	"github.com/grafana/grafana/pkg/services/accesscontrol/ossaccesscontrol"
 	"github.com/grafana/grafana/pkg/services/contexthandler/ctxkey"
 	"github.com/grafana/grafana/pkg/services/datasources"
 	"github.com/grafana/grafana/pkg/services/featuremgmt"
@@ -55,9 +55,9 @@ func setupTestServer(
 	}
 
 	var err error
-	acService, err := ossaccesscontrol.ProvideService(cfg, database.ProvideService(db), rr)
+	acService, err := acimpl.ProvideService(cfg, database.ProvideService(db), rr)
 	require.NoError(t, err)
-	ac := ossaccesscontrol.ProvideAccessControl(cfg, acService)
+	ac := acimpl.ProvideAccessControl(cfg, acService)
 
 	// build mux
 	m := web.New()
